part6/flux-architecture-and-redux: 6.8 anecdotes step 6
