+++ part6/redux-anecdotes/src/App.jsx
@@ -1,32 +1,11 @@
-import { useSelector, useDispatch } from "react-redux"
-
 import { AnecdoteForm } from "./components/AnecdoteForm"
-
-import { addVotes } from "./reducers/anecdoteReducer"
+import { AnecdoteList } from "./components/AnecdoteList"
 
 const App = () => {
-  const anecdotes = useSelector((state) => state.sort((a, b) => b.votes - a.votes))
-  const dispatch = useDispatch()
-
-  const vote = (id) => {
-    console.log("vote", id)
-    dispatch(addVotes(id))
-  }
-
   return (
     <div>
       <h2>Anecdotes</h2>
-
-      {anecdotes.map((anecdote) => (
-        <div key={anecdote.id}>
-          <div>{anecdote.content}</div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      ))}
-
+      <AnecdoteList />
       <AnecdoteForm />
     </div>
   )
